ファイルをサーバにアップする処理を別クラスへ分割
